setup project for win as well

+++ src/client.cpp
@@ -2,12 +2,13 @@
 
 #include "rpc_client.h"
 
+#include <cstdint>
 #include <memory>
 
 namespace mini_rpc {
 
 namespace session {
-constexpr uint BUFF_SIZE = 1024;
+constexpr uint16_t BUFF_SIZE = 1024;
 }
 
 Client::Client(std::string _endpoint) : rpc(std::make_unique<RpcClient>(_endpoint)) {}
